feat: sites endpoint refactor: move overview endpoint to common repo
