+++ pom.xml
@@ -17,12 +17,12 @@
     <dependency>
       <groupId>com.bigboxer23</groupId>
       <artifactId>solar-moon-common</artifactId>
-      <version>1.6.22</version>
+      <version>1.6.23</version>
     </dependency>
     <dependency>
       <groupId>com.bigboxer23</groupId>
       <artifactId>solar-moon-common</artifactId>
-      <version>1.6.22</version>
+      <version>1.6.23</version>
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
