--- src/main/java/com/bigboxer23/solar_moon/lambda/overview/SiteOverviewData.java
@@ -1,20 +1,0 @@
-package com.bigboxer23.solar_moon.lambda.overview;
-
-import lombok.Data;
-import org.opensearch.client.opensearch.core.SearchResponse;
-
-/** */
-@Data
-public class SiteOverviewData {
-	private SearchResponse totalAvg;
-
-	private SearchResponse total;
-
-	private SearchResponse avg;
-
-	private SearchResponse timeSeries;
-
-	private SearchResponse dailyEnergyConsumedTotal;
-
-	private double dailyEnergyConsumedAverage;
-}
